renames script.js to app.js
