+++ public/index.html
@@ -48,7 +48,7 @@
     <link rel="stylesheet" href="./css/main.css" />
     <link rel="stylesheet" type="text/css" href="css/main.css" />
     <script defer type="module">
-      import app from './js/script.js'
+      import app from './js/app.js'
       app.init('Check It Out Now')
     </script>
     <!--this is for trying out the code to see if it works, but I don't have access to app here-->
