+++ public/js/views/index.js
@@ -1,7 +1,7 @@
-import Controller from "./controller.js";
-import { kebabCase } from "./utilities/utilities.js";
+import app from '../app.js';
+import { kebabCase } from "../utilities/utilities.js";
 let viewIndex = 1;
-export class View {
+export default class View {
   // What the app looks like, what the user can see and do, User Interface
   constructor(element, content, anchor, id, classList) {
     const container = document.createElement(element);
@@ -517,56 +517,3 @@ export class View {
       //       ? "block"
       //       : "none";
       // });
-    
-
-
-class App {
-  constructor() {
-    this.tasks = [];
-    this.views = [];
-    this.view = new View(
-      "div",
-      " <!--Code injected by Amanda-->",
-      document.body,
-      "app",
-      "container"
-    );
-    this.controller = new Controller();
-  }
-  //Function to initialize app:
-  async init(title) {
-    await this.controller.loadData();
-    this.controller.init(title);
-    return this;
-  }
-}
-const app = new App();
-export default app;
-//Object constructor to create new tasks:
-export class Task {
-  constructor(name, group, category, frequency, days, calander) {
-    this.name = name;
-    this.group = group;
-    this.category = category;
-    this.frequency = frequency;
-    this.days = days;
-    this.calander = calander;
-    this.complete = false;
-    app.tasks.push(this);
-  }
-  read() {
-    this.complete = false;
-    return this;
-  }
-  //Function to create new tasks:
-  update(name, group, category, frequency, days, calander) {
-    this.name = name;
-    this.group = group;
-    this.category = category;
-    this.frequency = frequency;
-    this.days = days;
-    this.calander = calander;
-    this.complete = false;
-    return this;
-  }
-}
